+++ pyArango/query.py
@@ -113,13 +113,10 @@ class Query(object):
         return self
 
     def __getitem__(self, i):
-        "returns a ith result of the query."
+        "returns a ith result of the query. Raises IndexError if we reached the end of the current batch."
         if not self.rawResults and (not isinstance(self.result[i], (Edge, Document))):
             self._developDoc(i)
-        try:
-            return self.result[i]
-        except IndexError as e:
-            self.nextBatch()
+        return self.result[i]
 
     def __len__(self):
         """Returns the number of elements in the query results"""